Fix song truncation problem and calculated length
This first issue addressed here is an error which causes truncation of a
song if the end of a FLAC file is reached but rest of the mp3 is longer
than the size requested for read. This occurred rarely, and would result
in all future read requests failing, so that the last bit of the file
would be inaccessible.

The second issue is related, which is the correct calculation of the
file size ahead of time. This became necessary for the correct solution
to the above problem.