Add more build system optimizations
We do not need libtool, since we do not make any shared libraries.

We can use autoreconf to automatically run the programs we need to set up
the build system.

A new rule makes it easy to remove every generated file. This should not
be used by end-users. ``squeaky-clean: for when maintainer-clean is not
enough.''