Fix all warnings reported by new compiler flags
A number of unused variables are removed. In addition, libFLAC and
libid3tag disagree on whether strings should be signed, so we add an
explicit cast the remove the warning. This is safe because only strict
ASCII characters are used, which have the same value in either case.

In addition we fix a bug introduced a year and a half ago in fa943b6.
PRIu64 is not portable and does not include a percent symbol so we just
replace it with an equivalent string literal.