Update numbers of blocks in README
The same number of blocks is shown for the original and mp3 file list in
the README. The does not reflect the fix in a recent version of mp3fs,
so we correct the number with simulated values based on a 4096 byte disk
block size rerepresented as 1024 byte blocks. This is for example what
is done by default in recent versions of coreutils.