Make debug logging a runtime option
Debug logging is no longer done through a conditionally compiled macro.
Instead, it is a runtime option enabled with the same parameters (-d and
-odebug) as FUSE employs to turn on debugging to the console. In
addition, all messages are always recorded to the syslog facility. The
syntax of logging messages is slightly different, with mp3fs_debug
replacing DEBUG, and two new macros added for other types of messages.