Make the manpage generation reproducible.
Whilst working on the Reproducible Builds effort [0], we noticed that mp3fs
could not be built reproducibly as it uses a locale, timezone and
current-time-of-day header in the generated manpage.

This patch ensures that the output is consistent regardless of the build user's
location, or by the current time by utilising the SOURCE_DATE_EPOCH environment
variable.

  [0] https://reproducible-builds.org/
  [1] https://reproducible-builds.org/specs/source-date-epoch/