Rationalize mp3fs treatment of file sizes
When an encode is finished, the number of bytes returned by the encoder
might differ from the number calculated at the beginning. This change
ensures that the number of bytes returned always matches the result of a
stat operation. If the stat cache is enabled, we can return the number
of bytes according to the encoder; since the size is cached, this will
match a future stat. If it is disabled, we must make the file size match
what we computed, either by truncating or extending.

In either case, the ID3v1 tag must remain intact in the correct position
in order to remain valid, which this change also ensures.

This also removes the "actual size" logic from the encoder and the
stored size from the transcoder, both of which are no longer needed.
Instead, the buffer size is used as the file size.